refactor: unify streaming package to source-only Bundler mode
Adopt 2026 monorepo best practice: internal packages ship source
directly, no build step. All projects use moduleResolution: Bundler.

- Move packages/streaming → clients/shared/streaming
- Switch streaming + CLI tsconfig from NodeNext to Bundler
- Remove dist/ build, prepare hook, shared-build Make target
- Package exports point to src/index.ts directly
- Web uses transpilePackages for source transpilation
- CLI uses tsx which handles extension-less imports natively
