+++ clients/shared/streaming/package.json
@@ -3,8 +3,11 @@
   "version": "0.1.0",
   "description": "Shared LangGraph streaming types, constants, and utilities for Web and CLI",
   "type": "module",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
   "scripts": {
     "typecheck": "tsc --noEmit"
   },
