+++ clients/shared/streaming/src/constants.ts
