+++ clients/shared/streaming/src/types.ts
