+++ clients/shared/streaming/src/utils.ts
