+++ clients/web/tsconfig.json
@@ -29,6 +29,6 @@
     ".next/types/**/*.ts",
     ".next/dev/types/**/*.ts",
     "**/*.mts"
-  ],
+, "next.config.ts"  ],
   "exclude": ["node_modules"]
 }
